pyproject.toml
[tool.poetry]
name = "temple-price-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.9,<4"
"discord.py" = "^1.7.3"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
loguru = "^0.6.0"
web3 = "^7.14.0"

[tool.poetry.dev-dependencies]

[tool.pydocstyle]
ignore=["D100", "D101", "D103"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
